More cleanup. Still no changes to algorithm
