+++ python-sdk/nuscenes/eval/detection/render.py
@@ -1,0 +1,81 @@
+from typing import Dict
+
+import numpy as np
+from matplotlib import pyplot as plt
+from nuscenes.eval.detection.utils import boxes_to_sensor
+from nuscenes.nuscenes import NuScenes
+from nuscenes.utils.data_classes import LidarPointCloud
+from nuscenes.utils.geometry_utils import view_points
+
+
+def visualize_sample(nusc: NuScenes,
+                     sample_token: str,
+                     all_annotations: Dict,
+                     all_results: Dict,
+                     nsweeps: int = 1,
+                     conf_th: float = 0.15,
+                     eval_range: float = 40,
+                     verbose=True) -> None:
+    """
+    Visualizes a sample from BEV with annotations and detection results.
+    :param nusc: NuScenes object.
+    :param sample_token: The nuScenes sample token.
+    :param all_annotations: Maps each sample token to its annotations.
+    :param all_results: Maps each sample token to its results.
+    :param nsweeps: Number of sweeps used for lidar visualization.
+    :param conf_th: The confidence threshold used to filter negatives.
+    :param eval_range: Range in meters beyond which boxes are ignored.
+    :param verbose: Whether to print to stdout.
+    """
+
+    # Retrieve sensor & pose records.
+    sample_rec = nusc.get('sample', sample_token)
+    sd_record = nusc.get('sample_data', sample_rec['data']['LIDAR_TOP'])
+    cs_record = nusc.get('calibrated_sensor', sd_record['calibrated_sensor_token'])
+    pose_record = nusc.get('ego_pose', sd_record['ego_pose_token'])
+
+    # Get boxes.
+    boxes_gt_global = all_annotations[sample_token]
+    boxes_est_global = all_results[sample_token]
+
+    # Map GT boxes to lidar.
+    boxes_gt = boxes_to_sensor(boxes_gt_global, pose_record, cs_record)
+
+    # Map EST boxes to lidar.
+    boxes_est = boxes_to_sensor(boxes_est_global, pose_record, cs_record)
+
+    # Get point cloud in lidar frame.
+    pc, _ = LidarPointCloud.from_file_multisweep(nusc, sample_rec, 'LIDAR_TOP', 'LIDAR_TOP', nsweeps=nsweeps)
+
+    # Init axes.
+    _, ax = plt.subplots(1, 1, figsize=(9, 9))
+
+    # Show point cloud.
+    points = view_points(pc.points[:3, :], np.eye(4), normalize=False)
+    dists = np.sqrt(np.sum(pc.points[:2, :] ** 2, axis=0))
+    colors = np.minimum(1, dists / eval_range)
+    ax.scatter(points[0, :], points[1, :], c=colors, s=0.2)
+
+    # Show ego vehicle.
+    ax.plot(0, 0, 'x', color='black')
+
+    # Show GT boxes.
+    for box in boxes_gt:
+        box.render(ax, view=np.eye(4), colors=('g', 'g', 'g'), linewidth=2)
+
+    # Show EST boxes.
+    for box in boxes_est:
+        # Show only predictions with a high score.
+        if box.score >= conf_th:
+            box.render(ax, view=np.eye(4), colors=('b', 'b', 'b'), linewidth=1)
+
+    # Limit visible range.
+    axes_limit = eval_range + 3  # Slightly bigger to include boxes that extend beyond the range.
+    ax.set_xlim(-axes_limit, axes_limit)
+    ax.set_ylim(-axes_limit, axes_limit)
+
+    # Show plot.
+    if verbose:
+        print('Showing sample token %s' % sample_token)
+    plt.title(sample_token)
+    plt.show()
